+++ src/js/addon-background.js
@@ -1,4 +1,0 @@
-chrome.storage.local.get('background', function(res) {
-    if (res.background)
-        chrome.storage.local.remove('background');
-});